Issue #99 : withTouch() touchstart event was relying on direct access to the eventHandlers array within the Processing module to remove any existing mouse event handlers.
Instead of exposing the eventHandlers array directly, this change exposes a
method called detachEventHandlersByTypeMatch() which takes a regex of event
types, iterates through the list of existing event handlers, and detaches the
handlers that match.

The touchstart handler then calls detachEventHandlersByTypeMatch() with a regex
that matches the desired mouse events.