Update Math.js to fix Issue #244
Added a retry loop to random() so that it will never return its second argument (the "upper" bound) except in the truly degenerate case where it equals its first argument (the "lower" bound).  By avoiding > and < , the fix continues to support callers who disregard the notions of "lower" and "upper", e.g. random(10.2, -5).  The worst case probability of having to retry a new value of ir is 50%.