Update Math.js to fix Issue 256
The fix to this issue is the same as the fix to Issue 244: if the result is out of range, try again.  The function was restructured to invoke `internalRandomGenerator` at only one location after the function's parameters' values have been set.  The test for the degenerate case where `a===b` and the not-so-infinite retry loop remain intact.  The function's header comment was cleaned up and ill chosen words like "high", "low" and "top" were removed.  Similarly, neutral variable names replaced those with "min" and "max".